+++ R/utilities.R
@@ -1,4 +1,14 @@
 
+mydownload <- function(url, method, quiet = TRUE, ...) {
+    if (capabilities("libcurl")) {
+        dl <- tryCatch(utils::download.file(url, quiet = quiet, method = "libcurl", ...),
+                       error = function(e) NULL)
+    } else {
+        dl <- tryCatch(downloader::download(url, quiet = TRUE, method = method, ...),
+                       error = function(e) NULL)
+    }
+    return(dl)
+}